Add Renderer class & make helper for openGL error
