+++ src/MainAppController/common/VertexArray.cc
@@ -1,6 +1,7 @@
 #include "VertexArray.h"
 
-#include <Renderer.h>
+#include <openGLErrorHelpers.h>
+#include <openGLHeaders.h>
 
 VertexArray::VertexArray() { GLCall(glGenVertexArrays(1, &identifier_)); }
