+++ src/MainAppController/common/VertexArray.h
@@ -4,6 +4,7 @@
 #include <VertexBuffer.h>
 #include <VertexBufferLayout.h>
 
+class VertexBufferLayout;
 class VertexArray {
  private:
   unsigned int identifier_{0};
